[Asset] 무기들 추가
및 어드레서블 처리

+++ Assets/02. Scripts/UserData/UserData.cs
@@ -27,7 +27,7 @@ public class PreviewData
     public int slotNumber;
     public int achivementCount;
     public string time;
-    public double gold;
+    public long gold;
 }
 
 [Serializable]
